create bars class

+++ src/Array.cpp
@@ -15,7 +15,7 @@
 
 // Fill the array with normalized values 
 void Array::inflate() {
-    double bar_value_diff = (SCR_HEIGHT/2 + SCR_HEIGHT*0.1) / static_cast<double>(_num_elements);   // difference between each neighbouring value
+    double bar_value_diff = static_cast<double>(SCR_HEIGHT/2 + SCR_HEIGHT*0.1) / static_cast<double>(_num_elements);   // difference between each neighbouring value
 
     double i = 1.0;
